// apps/super-app/src/libs/sanity/sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export interface SanityImageAssetReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
}
// ...
export interface AiToolSectionImage {
  _type: "aiToolSectionImage";
  asset?: SanityImageAssetReference;
  media?: unknown;
  hotspot?: SanityImageHotspot;
  crop?: SanityImageCrop;
  alt?: string;
  title?: string;
}

export interface AiToolRichText {
  _type: "aiToolRichText";
  prefix?: string;
  main?: string;
  suffix?: string;
}

export interface AiSnippetSetReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "aiSnippetSet";
}

export interface AiToolBanner {
  _type: "aiToolBanner";
  heading?: AiToolRichText;
  description?: string;
  placeholder?: string;
  contentStyle?:
    | "default"
    | "translate"
    | "qa-cards"
    | "qa-simple"
    | "upload-file";
  allowSelectModel?: boolean;
  redirectLink?: string;
  promptSnippets?: ({
    _key: string;
  } & AiSnippetSetReference)[];
}

export interface AiToolPreview {
  _type: "aiToolPreview";
  image?: AiToolSectionImage;
  preview_title?: string;
  preview_description?: string;
}

export interface AiImageUpload {
  _id: string;
  _type: "aiImageUpload";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  image?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
}

export interface SanityImageCrop {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
}

export interface SanityImageHotspot {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
}
// ...
export interface AiToolFaqItemsByLocale {
  _type: "aiToolFaqItemsByLocale";
  en?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  zh?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  th?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  ar?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  es?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  ko?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  ja?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  hi?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
}

export interface AiToolFaqQaPair {
  _type: "aiToolFaqQaPair";
  question?: string;
  answer?: {
    children?: {
      marks?: string[];
      text?: string;
      _type: "span";
      _key: string;
    }[];
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: {
      href?: string;
      _type: "link";
      _key: string;
    }[];
    level?: number;
    _type: "block";
    _key: string;
  }[];
  answerPlain?: string;
}

export interface AiSeoReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "aiSeo";
}
// ...
export interface AiSeoLocaleContent {
  _type: "aiSeoLocaleContent";
  title?: string;
  description?: string;
  ogImage?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  ogImageAlt?: string;
}
// ...
export interface Slug {
  _type: "slug";
  current?: string;
  source?: string;
}
// ...
export interface BlogsReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "blogs";
}

export interface CategoryReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "category";
}

export interface BlogTagReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "blogTag";
}
// ...
export interface AiToolFAQsReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "aiToolFAQs";
}

export interface Pricing {
  _id: string;
  _type: "pricing";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  language?: "en" | "zh" | "th" | "ar" | "es" | "ko" | "ja" | "hi";
  name?: string;
  seo?: AiSeoReference;
  faq?: AiToolFAQsReference;
}

export interface AiToolSectionSchemaReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "aiToolSectionSchema";
}

export interface AiTool {
  _id: string;
  _type: "aiTool";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  groupId?:
    | "ai-tool-image"
    | "ai-tool-text"
    | "ai-tool-model"
    | "ai-tool-code"
    | "ai-tool-doc"
    | "ai-tool-study";
  language?: "en" | "zh" | "th" | "ar" | "es" | "ko" | "ja" | "hi";
  slug?: Slug;
  title?: string;
  preview?: AiToolPreview;
  linkAiTool?: {
    richText?: AiToolRichText;
    subTitle?: string;
    groupIds?: (
      | "ai-tool-image"
      | "ai-tool-text"
      | "ai-tool-model"
      | "ai-tool-code"
      | "ai-tool-doc"
      | "ai-tool-study"
    )[];
  };
  banner?: AiToolBanner;
  seo?: AiSeoReference;
  blogTags?: ({
    _key: string;
  } & BlogTagReference)[];
  blogs?: ({
    _key: string;
  } & BlogsReference)[];
  faq?: AiToolFAQsReference;
  sections?: ({
    _key: string;
  } & AiToolSectionSchemaReference)[];
}

export interface AiToolFAQs {
  _id: string;
  _type: "aiToolFAQs";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  itemsByLocale?: AiToolFaqItemsByLocale;
}

export interface AiSeo {
  _id: string;
  _type: "aiSeo";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  en?: AiSeoLocaleContent;
  zh?: AiSeoLocaleContent;
  th?: AiSeoLocaleContent;
  ar?: AiSeoLocaleContent;
  es?: AiSeoLocaleContent;
  ko?: AiSeoLocaleContent;
  ja?: AiSeoLocaleContent;
  hi?: AiSeoLocaleContent;
}